remove redundant function
