+++ examples/python/keras/reduce_sum.py
@@ -72,24 +72,9 @@ def test_reduce_sum3():
     epochs = 2
   )
 
-def get_configs():
-  import argparse,json
-  parser = argparse.ArgumentParser()
-  parser.add_argument(
-    "-config-file",
-    help="The path to a JSON file with the configs. If omitted, a sample model and configs will be used instead.",
-    type=str,
-    default=None,
-  )
-  args, unknown = parser.parse_known_args()
-  if args.config_file is not None:
-    with open(args.config_file) as f:
-      return json.load(f)
-  else:
-    return None
 
 if __name__ == "__main__":
-  configs = get_configs()
+  configs = ff.get_configs()
   ff.init_flexflow_runtime(configs)
   test_reduce_sum1()
   test_reduce_sum2()
