+++ examples/python/keras/seq_mnist_cnn.py
@@ -55,24 +55,9 @@ def top_level_task():
 
   model.fit(x_train, y_train, epochs=5, callbacks=[VerifyMetrics(ModelAccuracy.MNIST_CNN), EpochVerifyMetrics(ModelAccuracy.MNIST_CNN)])
 
-def get_configs():
-  import argparse,json
-  parser = argparse.ArgumentParser()
-  parser.add_argument(
-    "-config-file",
-    help="The path to a JSON file with the configs. If omitted, a sample model and configs will be used instead.",
-    type=str,
-    default=None,
-  )
-  args, unknown = parser.parse_known_args()
-  if args.config_file is not None:
-    with open(args.config_file) as f:
-      return json.load(f)
-  else:
-    return None
 
 if __name__ == "__main__":
   print("Sequential model, mnist cnn")
-  configs = get_configs()
+  configs = ff.get_configs()
   ff.init_flexflow_runtime(configs)
   top_level_task()
