+++ python/flexflow/config.py
@@ -56,3 +56,20 @@ _FF_DIR = os.path.dirname(os.path.realpath(__file__))
 
 def flexflow_dir():
     return _FF_DIR
+
+# Get runtime configs from the command line 
+def get_configs():
+  import argparse,json
+  parser = argparse.ArgumentParser()
+  parser.add_argument(
+    "-config-file",
+    help="The path to a JSON file with the configs. If omitted, a sample model and configs will be used instead.",
+    type=str,
+    default=None,
+  )
+  args, unknown = parser.parse_known_args()
+  if args.config_file is not None:
+    with open(args.config_file) as f:
+      return json.load(f)
+  else:
+    return None
